% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{glance-jwdid}
\alias{glance-jwdid}
\alias{glance.jwdid}
\alias{glance.jwdid_aggte}
\title{Model-level summaries for jwdid results}
\usage{
\method{glance}{jwdid}(x, ...)

\method{glance}{jwdid_aggte}(x, ...)
}
\arguments{
\item{x}{A \code{jwdid} or \code{jwdid_aggte} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A one-row data frame of model or aggregation diagnostics.
}
\description{
Model-level summaries for jwdid results
}
