% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggte.R
\name{jwdid_aggte}
\alias{jwdid_aggte}
\title{Aggregate treatment effects}
\usage{
jwdid_aggte(
  object,
  type = c("simple", "group", "calendar", "event"),
  weights = NULL,
  orestriction = NULL,
  over = NULL,
  over2 = NULL,
  window = NULL,
  cwindow = NULL,
  pretrend = FALSE,
  engine = c("auto", "analytic", "marginaleffects"),
  ...
)
}
\description{
Aggregate treatment effects
}
