% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy-jwdid}
\alias{tidy-jwdid}
\alias{tidy.jwdid}
\alias{tidy.jwdid_aggte}
\title{Tidy jwdid results}
\usage{
\method{tidy}{jwdid}(x, ...)

\method{tidy}{jwdid_aggte}(x, ...)
}
\arguments{
\item{x}{A \code{jwdid} or \code{jwdid_aggte} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A data frame containing one row per coefficient or aggregate.
}
\description{
Tidy jwdid results
}
