% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwdid.R
\name{jwdid}
\alias{jwdid}
\title{Estimate an extended two-way fixed-effects DiD model}
\usage{
jwdid(
  fml,
  data,
  ivar = NULL,
  tvar,
  gvar,
  trtvar = NULL,
  never = FALSE,
  anticipation = 0,
  hettype = "timecohort",
  group = FALSE,
  hettype_ll = NULL,
  hettype_ul = NULL,
  hettype_recode = NULL,
  hettype_evbase = NULL,
  xattvar = NULL,
  exogvar = NULL,
  xtvar = NULL,
  xgvar = NULL,
  fevar = NULL,
  xasis = FALSE,
  method = NULL,
  corr = FALSE,
  cre = FALSE,
  cluster = NULL,
  weights = NULL,
  ...
)
}
\description{
\code{anticipation = 0} corresponds to omitting Stata's \code{anticipation()} option.
R value \code{k} corresponds to Stata \code{anticipation(k + 1)}.
}
